requirements.txt
biopython==1.77
scipy==1.5.2
pybedtools==0.8.1
pandas==1.1.3
smart-open==3.0.0
shortuuid==1.0.1
